scrcpyx-mgr/pyproject.toml
[project]
name = "scrcpyx-mgr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiocache>=0.12.3",
    "grpcio>=1.76.0",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.76.0",
]

